+++ src/evohomeasync2/const.py
@@ -10,8 +10,8 @@ from typing import Final
 
 from _evohome.helpers import camel_to_snake
 
-from .schemas import API_STRFTIME as API_STRFTIME  # noqa: PLC0414
 from .schemas.const import (
+    API_STRFTIME as API_STRFTIME,  # noqa: PLC0414
     TccDayOfWeek,
     TccDhwState,
     TccFanMode,
