+++ src/evohomeasync2/schemas/__init__.py
@@ -1,434 +1,15 @@
 """Schema for the vendor's TCC v2 API.
 
-The API is undocumented, so the TypedDicts defined across the submodules of this
+The API is undocumented, so the TypedDicts (Tcc*T) defined across the submodules of this
 package (account.py, config.py, status.py, state.py, schedule.py) are the best known
 description of the vendor's schema and are the source of truth for static typing.
 
-The voluptuous schemas (TCC_*) are derived from those TypedDicts and serve a different
-purpose: runtime validation and coercion of the data returned by the API.
-
 JSON key names use camelCase throughout; the OAuth token endpoint is an accepted
 exception as it natively returns snake_case.
+
+JSON value StrEnums (Tcc*) use PascalCase, except for TccEntityType, which is
+camelCase, and is used for URL construction.
+
+The voluptuous schemas (TCC_*) are derived from those TypedDicts and serve a different
+purpose: runtime validation and coercion of the data returned by the API endpoints.
 """
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Final
-
-import voluptuous as vol
-
-from _evohome.helpers import noop
-
-from .account import (
-    TccErrorResponseT,
-    TccFailureResponseT,
-    TccOAuthTokenResponseT,
-    TccTaskResponseT,
-    TccUsrAccountResponseT,
-    factory_error_response,
-    factory_post_oauth_token,
-    factory_status_response,
-    factory_user_account,
-)
-from .config import (
-    TccAllowedSystemModeResponseT,
-    TccDhwConfigEntryT,
-    TccDhwConfigResponseT,
-    TccDhwScheduleCapabilitiesResponseT,
-    TccDhwStateCapabilitiesResponseT,
-    TccGwyConfigEntryT,
-    TccGwyConfigResponseT,
-    TccLocationOwnerInfoT,
-    TccLocConfigEntryT,
-    TccLocConfigResponseT,
-    TccTcsConfigEntryT,
-    TccTcsConfigResponseT,
-    TccTimeZoneInfoT,
-    TccZonConfigEntryT,
-    TccZonConfigResponseT,
-    TccZonScheduleCapabilitiesResponseT,
-    TccZonSetpointCapabilitiesResponseT,
-    factory_location_installation_info,
-    factory_user_locations_installation_info,
-)
-from .const import (
-    API_STRFTIME,
-    REGEX_DHW_ID,
-    REGEX_GATEWAY_ID,
-    REGEX_LOCATION_ID,
-    REGEX_SYSTEM_ID,
-    REGEX_ZONE_ID,
-    S2_ACTIVE_FAULTS,
-    S2_ALLOWED_FAN_MODES,
-    S2_ALLOWED_MODES,
-    S2_ALLOWED_SETPOINT_MODES,
-    S2_ALLOWED_STATES,
-    S2_ALLOWED_SYSTEM_MODES,
-    S2_CAN_BE_CHANGED,
-    S2_CAN_BE_PERMANENT,
-    S2_CAN_BE_TEMPORARY,
-    S2_CAN_CONTROL_COOL,
-    S2_CAN_CONTROL_HEAT,
-    S2_CITY,
-    S2_COOL_SETPOINT,
-    S2_COOL_SETPOINT_VALUE,
-    S2_COUNTRY,
-    S2_CRC,
-    S2_CURRENT_OFFSET_MINUTES,
-    S2_DAILY_SCHEDULES,
-    S2_DAY_OF_WEEK,
-    S2_DHW,
-    S2_DHW_ID,
-    S2_DHW_STATE,
-    S2_DHW_STATE_CAPABILITIES_RESPONSE,
-    S2_DISPLAY_NAME,
-    S2_DOMESTIC_HOT_WATER,
-    S2_FAN_MODE,
-    S2_FAN_STATUS,
-    S2_FAULT_TYPE,
-    S2_FIRSTNAME,
-    S2_GATEWAY,
-    S2_GATEWAY_ID,
-    S2_GATEWAY_INFO,
-    S2_GATEWAYS,
-    S2_HEAT_SETPOINT,
-    S2_HEAT_SETPOINT_VALUE,
-    S2_IS_AVAILABLE,
-    S2_IS_CANCELABLE,
-    S2_IS_CHANGEABLE,
-    S2_IS_PERMANENT,
-    S2_IS_WI_FI,
-    S2_LANGUAGE,
-    S2_LASTNAME,
-    S2_LOCATION,
-    S2_LOCATION_ID,
-    S2_LOCATION_INFO,
-    S2_LOCATION_OWNER,
-    S2_LOCATION_TYPE,
-    S2_MAC,
-    S2_MAX_COOL_SETPOINT,
-    S2_MAX_DURATION,
-    S2_MAX_HEAT_SETPOINT,
-    S2_MAX_SWITCHPOINTS_PER_DAY,
-    S2_MIN_COOL_SETPOINT,
-    S2_MIN_DURATION,
-    S2_MIN_HEAT_SETPOINT,
-    S2_MIN_SWITCHPOINTS_PER_DAY,
-    S2_MODE,
-    S2_MODEL_TYPE,
-    S2_NAME,
-    S2_OFFSET_MINUTES,
-    S2_PERMANENT,
-    S2_POSTCODE,
-    S2_SCHEDULE_CAPABILITIES,
-    S2_SCHEDULE_CAPABILITIES_RESPONSE,
-    S2_SETPOINT_CAPABILITIES,
-    S2_SETPOINT_DEADBAND,
-    S2_SETPOINT_MODE,
-    S2_SETPOINT_STATUS,
-    S2_SETPOINT_VALUE_RESOLUTION,
-    S2_SINCE,
-    S2_STATE,
-    S2_STATE_STATUS,
-    S2_STREET_ADDRESS,
-    S2_SUPPORTS_DAYLIGHT_SAVING,
-    S2_SWITCHPOINTS,
-    S2_SYSTEM_ID,
-    S2_SYSTEM_MODE,
-    S2_SYSTEM_MODE_STATUS,
-    S2_TARGET_COOL_TEMPERATURE,
-    S2_TARGET_HEAT_TEMPERATURE,
-    S2_TEMPERATURE,
-    S2_TEMPERATURE_CONTROL_SYSTEM,
-    S2_TEMPERATURE_CONTROL_SYSTEMS,
-    S2_TEMPERATURE_STATUS,
-    S2_TEMPERATURE_ZONE,
-    S2_TIME_OF_DAY,
-    S2_TIME_UNTIL,
-    S2_TIME_ZONE,
-    S2_TIME_ZONE_ID,
-    S2_TIMING_MODE,
-    S2_TIMING_RESOLUTION,
-    S2_UNTIL,
-    S2_UNTIL_TIME,
-    S2_USE_DAYLIGHT_SAVE_SWITCHING,
-    S2_USER_ACCOUNT,
-    S2_USER_ID,
-    S2_USERNAME,
-    S2_VACATION_HOLD_CAPABILITIES,
-    S2_VALUE_RESOLUTION,
-    S2_ZONE_ID,
-    S2_ZONE_TYPE,
-    S2_ZONES,
-    TccDayOfWeek,
-    TccDhwState,
-    TccEntityType,
-    TccFanMode,
-    TccFaultType,
-    TccLocationType,
-    TccSystemMode,
-    TccTcsModelType,
-    TccZoneMode,
-    TccZoneModelType,
-    TccZoneType,
-)
-from .helpers import Case
-from .schedule import (
-    TccDhwDailySchedulesT,
-    TccDhwDayOfWeekT,
-    TccDhwSwitchpointT,
-    TccZonDailySchedulesT,
-    TccZonDayOfWeekT,
-    TccZonSwitchpointT,
-    factory_dhw_schedule,
-    factory_zon_schedule,
-)
-from .state import TccSetDhwModeT, TccSetTcsModeT, TccSetZonModeT
-from .status import (
-    TccActiveFaultResponseT,
-    TccDhwStateStatusResponseT,
-    TccDhwStatusResponseT,
-    TccGwyStatusResponseT,
-    TccLocStatusResponseT,
-    TccSystemModeStatusResponseT,
-    TccTcsStatusResponseT,
-    TccTemperatureStatusResponseT,
-    TccZonSetpointStatusResponseT,
-    TccZonStatusResponseT,
-    factory_dhw_status,
-    factory_gwy_status,
-    factory_loc_status,
-    factory_tcs_status,
-    factory_zon_status,
-)
-
-__all__ = [
-    "API_STRFTIME",
-    "REGEX_DHW_ID",
-    "REGEX_GATEWAY_ID",
-    "REGEX_LOCATION_ID",
-    "REGEX_SYSTEM_ID",
-    "REGEX_ZONE_ID",
-    "S2_ACTIVE_FAULTS",
-    "S2_ALLOWED_FAN_MODES",
-    "S2_ALLOWED_MODES",
-    "S2_ALLOWED_SETPOINT_MODES",
-    "S2_ALLOWED_STATES",
-    "S2_ALLOWED_SYSTEM_MODES",
-    "S2_CAN_BE_CHANGED",
-    "S2_CAN_BE_PERMANENT",
-    "S2_CAN_BE_TEMPORARY",
-    "S2_CAN_CONTROL_COOL",
-    "S2_CAN_CONTROL_HEAT",
-    "S2_CITY",
-    "S2_COOL_SETPOINT",
-    "S2_COOL_SETPOINT_VALUE",
-    "S2_COUNTRY",
-    "S2_CRC",
-    "S2_CURRENT_OFFSET_MINUTES",
-    "S2_DAILY_SCHEDULES",
-    "S2_DAY_OF_WEEK",
-    "S2_DHW",
-    "S2_DHW_ID",
-    "S2_DHW_STATE",
-    "S2_DHW_STATE_CAPABILITIES_RESPONSE",
-    "S2_DISPLAY_NAME",
-    "S2_DOMESTIC_HOT_WATER",
-    "S2_FAN_MODE",
-    "S2_FAN_STATUS",
-    "S2_FAULT_TYPE",
-    "S2_FIRSTNAME",
-    "S2_GATEWAY",
-    "S2_GATEWAYS",
-    "S2_GATEWAY_ID",
-    "S2_GATEWAY_INFO",
-    "S2_HEAT_SETPOINT",
-    "S2_HEAT_SETPOINT_VALUE",
-    "S2_IS_AVAILABLE",
-    "S2_IS_CANCELABLE",
-    "S2_IS_CHANGEABLE",
-    "S2_IS_PERMANENT",
-    "S2_IS_WI_FI",
-    "S2_LANGUAGE",
-    "S2_LASTNAME",
-    "S2_LOCATION",
-    "S2_LOCATION_ID",
-    "S2_LOCATION_INFO",
-    "S2_LOCATION_OWNER",
-    "S2_LOCATION_TYPE",
-    "S2_MAC",
-    "S2_MAX_COOL_SETPOINT",
-    "S2_MAX_DURATION",
-    "S2_MAX_HEAT_SETPOINT",
-    "S2_MAX_SWITCHPOINTS_PER_DAY",
-    "S2_MIN_COOL_SETPOINT",
-    "S2_MIN_DURATION",
-    "S2_MIN_HEAT_SETPOINT",
-    "S2_MIN_SWITCHPOINTS_PER_DAY",
-    "S2_MODE",
-    "S2_MODEL_TYPE",
-    "S2_NAME",
-    "S2_OFFSET_MINUTES",
-    "S2_PERMANENT",
-    "S2_POSTCODE",
-    "S2_SCHEDULE_CAPABILITIES",
-    "S2_SCHEDULE_CAPABILITIES_RESPONSE",
-    "S2_SETPOINT_CAPABILITIES",
-    "S2_SETPOINT_DEADBAND",
-    "S2_SETPOINT_MODE",
-    "S2_SETPOINT_STATUS",
-    "S2_SETPOINT_VALUE_RESOLUTION",
-    "S2_SINCE",
-    "S2_STATE",
-    "S2_STATE_STATUS",
-    "S2_STREET_ADDRESS",
-    "S2_SUPPORTS_DAYLIGHT_SAVING",
-    "S2_SWITCHPOINTS",
-    "S2_SYSTEM_ID",
-    "S2_SYSTEM_MODE",
-    "S2_SYSTEM_MODE_STATUS",
-    "S2_TARGET_COOL_TEMPERATURE",
-    "S2_TARGET_HEAT_TEMPERATURE",
-    "S2_TEMPERATURE",
-    "S2_TEMPERATURE_CONTROL_SYSTEM",
-    "S2_TEMPERATURE_CONTROL_SYSTEMS",
-    "S2_TEMPERATURE_STATUS",
-    "S2_TEMPERATURE_ZONE",
-    "S2_TIME_OF_DAY",
-    "S2_TIME_UNTIL",
-    "S2_TIME_ZONE",
-    "S2_TIME_ZONE_ID",
-    "S2_TIMING_MODE",
-    "S2_TIMING_RESOLUTION",
-    "S2_UNTIL",
-    "S2_UNTIL_TIME",
-    "S2_USERNAME",
-    "S2_USER_ACCOUNT",
-    "S2_USER_ID",
-    "S2_USE_DAYLIGHT_SAVE_SWITCHING",
-    "S2_VACATION_HOLD_CAPABILITIES",
-    "S2_VALUE_RESOLUTION",
-    "S2_ZONES",
-    "S2_ZONE_ID",
-    "S2_ZONE_TYPE",
-    "Case",
-    "TccActiveFaultResponseT",
-    "TccAllowedSystemModeResponseT",
-    "TccDayOfWeek",
-    "TccDhwConfigEntryT",
-    "TccDhwConfigResponseT",
-    "TccDhwDailySchedulesT",
-    "TccDhwDayOfWeekT",
-    "TccDhwScheduleCapabilitiesResponseT",
-    "TccDhwState",
-    "TccDhwStateCapabilitiesResponseT",
-    "TccDhwStateStatusResponseT",
-    "TccDhwStatusResponseT",
-    "TccDhwSwitchpointT",
-    "TccEntityType",
-    "TccErrorResponseT",
-    "TccFailureResponseT",
-    "TccFanMode",
-    "TccFaultType",
-    "TccGwyConfigEntryT",
-    "TccGwyConfigResponseT",
-    "TccGwyStatusResponseT",
-    "TccLocConfigEntryT",
-    "TccLocConfigResponseT",
-    "TccLocStatusResponseT",
-    "TccLocationOwnerInfoT",
-    "TccLocationType",
-    "TccOAuthTokenResponseT",
-    "TccSetDhwModeT",
-    "TccSetTcsModeT",
-    "TccSetZonModeT",
-    "TccSystemMode",
-    "TccSystemModeStatusResponseT",
-    "TccTaskResponseT",
-    "TccTcsConfigEntryT",
-    "TccTcsConfigResponseT",
-    "TccTcsModelType",
-    "TccTcsStatusResponseT",
-    "TccTemperatureStatusResponseT",
-    "TccTimeZoneInfoT",
-    "TccUsrAccountResponseT",
-    "TccZonConfigEntryT",
-    "TccZonConfigResponseT",
-    "TccZonDailySchedulesT",
-    "TccZonDayOfWeekT",
-    "TccZonScheduleCapabilitiesResponseT",
-    "TccZonSetpointCapabilitiesResponseT",
-    "TccZonSetpointStatusResponseT",
-    "TccZonStatusResponseT",
-    "TccZonSwitchpointT",
-    "TccZoneMode",
-    "TccZoneModelType",
-    "TccZoneType",
-]
-
-if TYPE_CHECKING:
-    from collections.abc import Callable
-
-#
-# HTTP GET/PUT & POST schemas are camelCase, not snake_case...
-
-TCC_ERROR_RESPONSE: Final = factory_error_response()
-TCC_STATUS_RESPONSE: Final = factory_status_response()
-
-# POST /Auth/OAuth/Token  # TODO: add this
-TCC_POST_OAUTH_TOKEN: Final = factory_post_oauth_token()
-
-# GET /userAccount
-TCC_GET_USR_ACCOUNT: Final = factory_user_account()
-
-# GET /location/installationInfo?userId={usr_id}&includeTemperatureControlSystems=True
-TCC_GET_USR_LOCATIONS: Final = factory_user_locations_installation_info()
-
-# GET /location/{loc_id}/installationInfo?includeTemperatureControlSystems=True
-TCC_GET_LOC_INSTALLATION_INFO: Final = factory_location_installation_info()
-
-# GET /location/{loc_id}/status?includeTemperatureControlSystems=True
-TCC_GET_LOC_STATUS: Final = factory_loc_status()
-
-# GET /gateway/{gwy_id}/status...
-TCC_GET_GWY_STATUS: Final = factory_gwy_status()
-
-# GET /temperatureControlSystem/{tcs_id}/status
-TCC_GET_TCS_STATUS: Final = factory_tcs_status()
-
-# GET /domesticHotWater/{dhw_id}/status
-TCC_GET_DHW_STATUS: Final = factory_dhw_status()
-
-# GET /temperatureZone/{zone_id}/heatSetpoint
-# TODO:
-
-# GET /temperatureZone/{zone_id}/status
-TCC_GET_ZON_STATUS: Final = factory_zon_status()
-
-# GET /domesticHotWater/{dhw_id}/schedule
-TCC_GET_DHW_SCHEDULE: Final = factory_dhw_schedule()
-
-# PUT /domesticHotWater/{dhw_id}/schedule
-TCC_PUT_DHW_SCHEDULE: Final = TCC_GET_DHW_SCHEDULE
-
-# GET /temperatureZone/{zone_id}/schedule
-TCC_GET_ZON_SCHEDULE: Final = factory_zon_schedule()
-
-# PUT /temperatureZone/{zone_id}/schedule
-TCC_PUT_ZON_SCHEDULE: Final = TCC_GET_ZON_SCHEDULE
-
-
-# for convenience...
-def factory_get_schedule(_: Callable[[str], str] = noop) -> vol.Schema:
-    """Factory for the schedule schema."""
-
-    from . import TCC_GET_DHW_SCHEDULE, TCC_GET_ZON_SCHEDULE  # noqa: PLC0415
-
-    return vol.Schema(
-        vol.Any(TCC_GET_DHW_SCHEDULE, TCC_GET_ZON_SCHEDULE),
-        extra=vol.PREVENT_EXTRA,
-    )
-
-
-TCC_GET_SCHEDULE: Final = factory_get_schedule()
-TCC_PUT_SCHEDULE: Final = TCC_GET_SCHEDULE
