+++ tests/tests/test_schemas_0.py
@@ -9,7 +9,8 @@ import pytest
 
 from _evohome.helpers import camel_to_snake, convert_keys_to_snake_case
 from evohomeasync2 import Location
-from evohomeasync2.schemas import (
+from evohomeasync2.schemas.config import factory_tcs, factory_time_zone
+from evohomeasync2.schemas.const import (
     S2_GATEWAY_ID,
     S2_GATEWAY_INFO,
     S2_GATEWAYS,
@@ -20,7 +21,6 @@ from evohomeasync2.schemas import (
     S2_TIME_ZONE,
     S2_USE_DAYLIGHT_SAVE_SWITCHING,
 )
-from evohomeasync2.schemas.config import factory_tcs, factory_time_zone
 from evohomeasync2.schemas.status import factory_loc_status
 
 from .conftest import ClientStub
