+++ tests/tests_rf/test_v2_urls_auth.py
@@ -16,14 +16,20 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 
-import evohomeasync2 as evo2
-from evohomeasync2 import schemas
 from evohomeasync2.const import API_STRFTIME, SystemMode, ZoneMode
+from evohomeasync2.schemas.account import TCC_GET_USR_ACCOUNT
+from evohomeasync2.schemas.config import TCC_GET_USR_LOCATIONS
+from evohomeasync2.schemas.status import (
+    TCC_GET_LOC_STATUS,
+    TCC_GET_TCS_STATUS,
+    TCC_GET_ZON_STATUS,
+)
 from tests.const import _DBG_USE_REAL_AIOHTTP
 
 from .common import should_fail_v2, should_work_v2, skipif_auth_failed
 
 if TYPE_CHECKING:
+    import evohomeasync2 as evo2
     from tests.conftest import EvohomeClientV2
 
 
@@ -35,9 +41,7 @@ async def _test_usr_account(evo: EvohomeClientV2) -> None:
 
     # STEP 1:
     url = "userAccount"
-    _ = await should_work_v2(
-        evo.auth, HTTPMethod.GET, url, schema=schemas.TCC_GET_USR_ACCOUNT
-    )
+    _ = await should_work_v2(evo.auth, HTTPMethod.GET, url, schema=TCC_GET_USR_ACCOUNT)
     # {
     #     'userId': '2263181',
     #     'username': 'nobody@nowhere.com',
@@ -96,7 +100,7 @@ async def _test_user_locations(evo: EvohomeClientV2) -> None:
     #
     url += "&includeTemperatureControlSystems=True"
     _ = await should_work_v2(
-        evo.auth, HTTPMethod.GET, url, schema=schemas.TCC_GET_USR_LOCATIONS
+        evo.auth, HTTPMethod.GET, url, schema=TCC_GET_USR_LOCATIONS
     )
 
     #
@@ -143,9 +147,7 @@ async def _test_loc_status(evo: EvohomeClientV2) -> None:
     )
 
     url += "?includeTemperatureControlSystems=True"
-    _ = await should_work_v2(
-        evo.auth, HTTPMethod.GET, url, schema=schemas.TCC_GET_LOC_STATUS
-    )
+    _ = await should_work_v2(evo.auth, HTTPMethod.GET, url, schema=TCC_GET_LOC_STATUS)
     _ = await should_fail_v2(
         evo.auth, HTTPMethod.PUT, url, status=HTTPStatus.METHOD_NOT_ALLOWED
     )
@@ -197,7 +199,7 @@ async def _test_tcs_status(evo: EvohomeClientV2) -> None:
     url = f"{tcs._TCC_TYPE}/{tcs.id}/status"
 
     old_status: dict[str, Any] = await should_work_v2(
-        evo.auth, HTTPMethod.GET, url, schema=schemas.TCC_GET_TCS_STATUS
+        evo.auth, HTTPMethod.GET, url, schema=TCC_GET_TCS_STATUS
     )  # type: ignore[assignment]
     # {
     #      'systemId': '1234567',
@@ -322,9 +324,7 @@ async def _test_zone_status(evo: EvohomeClientV2) -> None:
 
     #
     url = f"{zone._TCC_TYPE}/{zone.id}/status"
-    _ = await should_work_v2(
-        evo.auth, HTTPMethod.GET, url, schema=schemas.TCC_GET_ZON_STATUS
-    )
+    _ = await should_work_v2(evo.auth, HTTPMethod.GET, url, schema=TCC_GET_ZON_STATUS)
     # {
     #     'zoneId': '3432576',
     #     'temperatureStatus': {'temperature': 25.5, 'isAvailable': True},
